+++ tests/Commands/CommandsTest.lua
@@ -6,10 +6,12 @@ TestCommand = BaseTestClass:new()
         local command = __:new('Command')
         
         command
+            :setArgsValidator('test-args-validator')
             :setCallback('test-callback')
             :setDescription('test-description')
             :setOperation('test-operation')
 
+        lu.assertEquals('test-args-validator', command.argsValidator)
         lu.assertEquals('test-callback', command.callback)
         lu.assertEquals('test-description', command.description)
         lu.assertEquals('test-operation', command.operation)
@@ -35,5 +37,38 @@ TestCommand = BaseTestClass:new()
         local command = __:new('Command')
 
         lu.assertNotIsNil(command)
+    end
+
+    -- @covers Command:validateArgs()
+    function TestCommand:testValidateArgs()
+        local function execution(validator, expectedOutput)
+            local command = __:new('Command')
+
+            command:setArgsValidator(validator)
+
+            lu.assertEquals(expectedOutput, command:validateArgs())
+        end
+
+        -- no validator
+        execution(nil, 'valid')
+
+        -- validator returning invalid
+        execution(function() return 'invalid' end, 'invalid')
+    end
+
+    -- @covers Command:validateArgs()
+    function TestCommand:testValidateArgsWithMultipleParameters()
+        local command = __:new('Command')
+
+        -- makes sure the arguments are being passed to the validator
+        command:setArgsValidator(function(arg1, arg2)
+            command.arg1 = arg1
+            command.arg2 = arg2
+        end)
+
+        command:validateArgs('arg1', 'arg2')
+
+        lu.assertEquals('arg1', command.arg1)
+        lu.assertEquals('arg2', command.arg2)
     end
 -- end of TestCommand