+++ documentation/docs/changelog.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 2024.06.18 - version 1.5.0
+
+* Add a level property to the Player model and update the current player instance automatically when the player levels up
+* Allow commands to register
+[argument validators](resources/commands/command#validating-arguments)
+* Default to the
+[help operation](resources/commands/commands-handler#the-help-operation)
+when it's missing in the addon main slash command
+* Watch for the [PLAYER_LEVEL_UP](resources/facades/events#player_level_up) 
+event and forward it to the library listeners
+
 ## 2024.06.11 - version 1.4.0
 
 * Add a factory to create item instances from multiple sources